refactor(generator): change default `meta` path
